Create Packer templates and provisioners
Creates the initial Packer templates for creating VMware Fusion virtual
machines and DigitalOcean snapshots, as well as a script provisioner for
creating a basic, updated, ansible-ready machine with admin and ansible
accounts.

Also creates a template file for storing secret variables, which should
be copied and added to .gitignore to avoid accidently committing secrets
(passwords, API keys, etc.) to git.